requirements.txt
absl-py==0.11.0
astunparse==1.6.3
bert-extractive-summarizer==0.6.1
blis==0.7.4
cachetools==4.2.0
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
filelock==3.0.12
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
torch==1.7.1
Markdown==3.3.3
MarkupSafe==1.1.1
murmurhash==1.0.5
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
Pillow==8.0.1
plac==1.1.3
preshed==3.0.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
spacy==2.3.5
srsly==1.0.5

termcolor==1.1.0
thinc==7.4.5
threadpoolctl==2.1.0
tokenizers==0.9.4
tqdm==4.55.1
transformers==4.1.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
Werkzeug==1.0.1
wrapt==1.12.1
